+++ Src/Swagger/Extensions.cs
@@ -1,11 +1,9 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using NJsonSchema;
 using NJsonSchema.Generation;
 using NSwag;
 using NSwag.AspNetCore;
 using NSwag.Generation;
 using NSwag.Generation.AspNetCore;
-using NSwag.Generation.Processors.Contexts;
 using NSwag.Generation.Processors.Security;
 using System.Reflection;
 using System.Runtime.CompilerServices;
@@ -31,19 +29,25 @@ public static class Extensions
     /// enable support for FastEndpoints in swagger
     /// </summary>
     /// <param name="tagIndex">the index of the route path segment to use for tagging/grouping endpoints</param>
+    /// <param name="minEndpointVersion">endpoints lower than this vesion will not be included in the swagger doc</param>
     /// <param name="maxEndpointVersion">endpoints greater than this version will not be included in the swagger doc</param>
     /// <param name="shortSchemaNames">set to true to make schema names just the name of the class instead of full type name</param>
     /// <param name="removeEmptySchemas">
     /// set to true for removing empty schemas from the swagger document.
     /// <para>WARNING: enabling this also flattens the inheritance hierachy of the schmemas.</para>
     /// </param>
-    public static void EnableFastEndpoints(this AspNetCoreOpenApiDocumentGeneratorSettings settings, int tagIndex, int maxEndpointVersion, bool shortSchemaNames, bool removeEmptySchemas)
+    public static void EnableFastEndpoints(this AspNetCoreOpenApiDocumentGeneratorSettings settings,
+        int tagIndex,
+        int minEndpointVersion,
+        int maxEndpointVersion,
+        bool shortSchemaNames,
+        bool removeEmptySchemas)
     {
         settings.Title = AppDomain.CurrentDomain.FriendlyName;
         settings.SchemaNameGenerator = new SchemaNameGenerator(shortSchemaNames);
         settings.SchemaProcessors.Add(new ValidationSchemaProcessor());
         settings.OperationProcessors.Add(new OperationProcessor(tagIndex, removeEmptySchemas));
-        settings.DocumentProcessors.Add(new DocumentProcessor(maxEndpointVersion));
+        settings.DocumentProcessors.Add(new DocumentProcessor(minEndpointVersion, maxEndpointVersion));
     }
 
     /// <summary>
@@ -67,20 +71,24 @@ public static class Extensions
     /// <param name="serializerSettings">json serializer options</param>
     /// <param name="addJWTBearerAuth">set to false to disable auto addition of jwt bearer auth support</param>
     /// <param name="tagIndex">the index of the route path segment to use for tagging/grouping endpoints</param>
+    /// <param name="minEndpointVersion">endpoints lower than this vesion will not be included in the swagger doc</param>
     /// <param name="maxEndpointVersion">endpoints greater than this version will not be included in the swagger doc</param>
     /// <param name="shortSchemaNames">set to true if you'd like schema names to be the class name intead of the full name</param>
     /// <param name="removeEmptySchemas">
     /// set to true for removing empty schemas from the swagger document.
     /// <para>WARNING: enabling this also flattens the inheritance hierachy of the schmemas.</para>
     /// </param>
+    /// <param name="excludeNonFastEndpoints">if set to true, only FastEndpoints will show up in the swagger doc</param>
     public static IServiceCollection AddSwaggerDoc(this IServiceCollection services,
         Action<AspNetCoreOpenApiDocumentGeneratorSettings>? settings = null,
         Action<JsonSerializerOptions>? serializerSettings = null,
         bool addJWTBearerAuth = true,
         int tagIndex = 1,
+        int minEndpointVersion = 0,
         int maxEndpointVersion = 0,
         bool shortSchemaNames = false,
-        bool removeEmptySchemas = false)
+        bool removeEmptySchemas = false,
+        bool excludeNonFastEndpoints = false)
     {
         services.AddEndpointsApiExplorer();
         services.AddOpenApiDocument(s =>
@@ -89,7 +97,8 @@ public static class Extensions
             SelectedJsonNamingPolicy = stjOpts.PropertyNamingPolicy;
             serializerSettings?.Invoke(stjOpts);
             s.SerializerSettings = SystemTextJsonUtilities.ConvertJsonOptionsToNewtonsoftSettings(stjOpts);
-            s.EnableFastEndpoints(tagIndex, maxEndpointVersion, shortSchemaNames, removeEmptySchemas);
+            s.EnableFastEndpoints(tagIndex, minEndpointVersion, maxEndpointVersion, shortSchemaNames, removeEmptySchemas);
+            if (excludeNonFastEndpoints) s.OperationProcessors.Insert(0, new FastEndpointsFilter());
             if (addJWTBearerAuth) s.EnableJWTBearerAuth();
             settings?.Invoke(s);
             s.FlattenInheritanceHierarchy = removeEmptySchemas;