+++ Web/Program.cs
@@ -43,6 +43,14 @@ builder.Services
         s.DocumentName = "Release 2.0";
         s.Title = "FastEndpoints Sandbox";
         s.Version = "v2.0";
+    })
+
+    //only ver3 & only FastEndpoints
+    .AddSwaggerDoc(minEndpointVersion: 3, maxEndpointVersion: 3, excludeNonFastEndpoints: true, settings: s =>
+    {
+        s.DocumentName = "Release 3.0";
+        s.Title = "FastEndpoints Sandbox ver3 only";
+        s.Version = "v3.0";
     });
 
 var app = builder.Build();